+++ pyuarm/util.py
@@ -16,17 +16,20 @@ def init_logger(debug):
         logger.setLevel(logging.DEBUG)
     else:
         logger.setLevel(logging.INFO)
-    # formatter = logging.Formatter('%(name)s - %(levelname)s - %(message)s')
-    # #
     ch = logging.StreamHandler()
-    # # # if debug:
-    # # #     ch.setLevel(logging.DEBUG)
-    # # # else:
-    # # #     ch.setLevel(logging.INFO)
-    # ch.setFormatter(formatter)
     logger.addHandler(ch)
     logger.info('pyuarm version: ' + __version__)
 
+def set_debug(debug):
+    """
+    Dynamically changing log level
+    :param debug:
+    :return:
+    """
+    if debug:
+        logging.getLogger('pyuarm').setLevel(logging.DEBUG)
+    else:
+        logging.getLogger('pyuarm').setLevel(logging.INFO)
 
 def printf(msg, type=INFO):
     """